feat: complete migration from IPFS logging to zerolog
- Updated all logger calls in block package to use zerolog fluent API
- Fixed logger calls in node, pkg/cmd, and da packages
- Migrated test files to use zerolog.Nop() for testing
- Updated all go.mod files with proper dependencies
- Removed unused IPFS log imports
- Fixed type conversions for zerolog methods (Bytes vs Str)

All builds and tests now pass successfully with zerolog.

+++ block/manager.go
@@ -595,7 +595,7 @@ func (m *Manager) publishBlockInternal(ctx context.Context) error {
 	}
 
 	if m.config.Node.MaxPendingHeadersAndData != 0 && (m.pendingHeaders.numPendingHeaders() >= m.config.Node.MaxPendingHeadersAndData || m.pendingData.numPendingData() >= m.config.Node.MaxPendingHeadersAndData) {
-		m.logger.Warn().Int("pending_headers", m.pendingHeaders.numPendingHeaders()).Int("pending_data", m.pendingData.numPendingData()).Int("limit", m.config.Node.MaxPendingHeadersAndData).Msg("refusing to create block: pending headers or data reached limit")
+		m.logger.Warn().Uint64("pending_headers", m.pendingHeaders.numPendingHeaders()).Uint64("pending_data", m.pendingData.numPendingData()).Uint64("limit", m.config.Node.MaxPendingHeadersAndData).Msg("refusing to create block: pending headers or data reached limit")
 		return nil
 	}
